create centralized config service
