+++ test/unit/core/infrastructure/config/log/sentry-bootstrap.spec.ts
@@ -1,4 +1,3 @@
-import * as Sentry from '@sentry/nestjs';
 
 jest.mock('@sentry/nestjs', () => ({
     init: jest.fn(),
